refactor(examples): reorganize by language and simplify justfile
Address PR review comments about justfile complexity and directory structure.

Changes:

1. **Directory Restructure:**
   - Created `typescript/` for all TS/JS examples
   - Created `rust/` for all Rust examples
   - Moved examples into language-specific subdirectories
   - All examples now follow consistent folder structure

2. **Simplify Justfile:**
   - Replace 5+ specific `run-test-*` commands with 2 generic ones:
     * `run-test-ts <example> <test_dir> <runtime> [mode]`
     * `run-test-rust <example> <test_dir> [ws_url] [seed]`
   - Updated `npm-install` to work in typescript/ directory
   - Updated `run-authorize-and-store` to use generic commands
   - Updated `run-all-tests` to use generic commands
   - Changed mode names: "ws" -> "papi", "smoldot" remains

3. **Update Import Paths:**
   - Updated all TypeScript files to use '../' for shared modules
   - Fixed authorize-preimage-and-store hardcoded localhost URIs

4. **Documentation:**
   - Completely rewrote `examples/README.md` with new structure
   - Created `typescript/README.md` for TS-specific docs
   - Created `rust/README.md` for Rust-specific docs
   - Clear examples of new command usage

Benefits:
- Scalable: Adding examples doesn't require new justfile commands
- Consistent: All examples follow folder pattern
- Clear: Language separation obvious at a glance
- Maintainable: Each language manages its own dependencies

Migration:
- Old: `just run-test-authorize-and-store <test_dir> <runtime> ws`
- New: `just run-test-ts authorize-and-store <test_dir> <runtime> papi`
