+++ examples/rust/authorize-and-store/.gitignore
