+++ examples/rust/authorize-and-store/Cargo.toml
