+++ examples/rust/authorize-and-store/IMPROVED_CONFIG.md
