+++ examples/rust/authorize-and-store/METADATA.md
