+++ examples/rust/authorize-and-store/README.md
