+++ examples/rust/authorize-and-store/fetch_metadata.sh
