+++ examples/rust/authorize-and-store/src/main.rs
