+++ examples/typescript/api.js
