+++ examples/typescript/cid_dag_metadata.js
