+++ examples/typescript/common.js
