+++ examples/typescript/package.json
